[windows] fix crash detection parsing problems found in testing. (#20181)
* [windows] fix crash detection parsing problems found in testing.

Testing found two problems, based on different environment and crash
types.

The first is that we may not get an offending module.  Accept that,
and add a test for that type of crash.

Also, some crash outputs don't call out the bugcheck code, but it can
be parsed out from the call stack.  in that case, parse the output so
we can still retrieve the bugcheck code

* lint

* For internal check, add check for unknown offending module.
If so, don't report (but do log a warning).

Change text of external message to reflect that offender couldn't
be determined.