pkg/trace/api: fix container tagging over UDS (#20688)
We introduced a new connection type in #17917 which caused our special
case for UDS to fail, resulting in the loss of container tags for users
using UDS container tagging on cgroup v2 hosts.

This commit correctly checks for the new connection type and pulls the
UDS credentials out of it.

Cherry-pick of #20641